+++ lib/gamefic/query/text.rb
@@ -7,9 +7,9 @@ module Gamefic
     class Text < Base
       # @param argument [String, Regexp]
       # @param name [String, nil]
-      def initialize argument = /.*/, name: nil
-        super argument, name: name
-        validate
+      def initialize argument = /.*/, name: self.class.name
+        super(argument, name: name)
+        validate_argument
       end
 
       def argument
@@ -51,7 +51,7 @@ module Gamefic
         end
       end
 
-      def validate
+      def validate_argument
         return if argument.is_a?(String) || argument.is_a?(Regexp)
 
         raise ArgumentError, 'Invalid text query argument'